fix(permissions): circular dependency
