+++ packages/core/permissions/src/engine/hooks.ts
@@ -1,10 +1,9 @@
 import { cloneDeep, has, isArray } from 'lodash/fp';
 import { hooks } from '@strapi/utils';
-// eslint-disable-next-line node/no-extraneous-import
-import type { Permissions as PermissionsTypes } from '@strapi/types';
 
 import * as domain from '../domain';
 import type { Permission } from '../domain/permission';
+import { PermissionRule } from '../types';
 
 export interface PermissionEngineHooks {
   'before-format::validate.permission': ReturnType<typeof hooks.createAsyncBailHook>;
@@ -52,7 +51,7 @@ const createBeforeEvaluateContext = (permission: Permission) => ({
 });
 
 interface WillRegisterContextParams {
-  permission: PermissionsTypes.PermissionRule;
+  permission: PermissionRule;
   options: Record<string, unknown>;
 }
