+++ packages/core/permissions/src/engine/index.ts
@@ -2,8 +2,6 @@ import _ from 'lodash/fp';
 import qs from 'qs';
 import { Ability } from '@casl/ability';
 import { providerFactory } from '@strapi/utils';
-// eslint-disable-next-line import/no-extraneous-dependencies, node/no-extraneous-import
-import { Permissions as PermissionsTypes } from '@strapi/types';
 
 import {
   createEngineHooks,
@@ -15,6 +13,7 @@ import type { PermissionEngineHooks, HookName } from './hooks';
 
 import * as abilities from './abilities';
 import { Permission } from '../domain/permission';
+import { PermissionRule } from '../types';
 
 export { abilities };
 
@@ -30,9 +29,9 @@ export interface Engine {
   on(hook: HookName, handler: (...args: any[]) => any): Engine;
   generateAbility(permissions: Permission[], options?: object): Promise<Ability>;
   createRegisterFunction(
-    can: (permission: PermissionsTypes.PermissionRule) => unknown,
+    can: (permission: PermissionRule) => unknown,
     options: Record<string, unknown>
-  ): (permission: PermissionsTypes.PermissionRule) => Promise<unknown>;
+  ): (permission: PermissionRule) => Promise<unknown>;
 }
 
 export interface EngineParams {
@@ -42,7 +41,7 @@ export interface EngineParams {
 
 interface EvaluateParams {
   options: Record<string, unknown>;
-  register: (permission: PermissionsTypes.PermissionRule) => Promise<unknown>;
+  register: (permission: PermissionRule) => Promise<unknown>;
   permission: Permission;
 }
 
@@ -179,7 +178,7 @@ const newEngine = (params: EngineParams): Engine => {
      * used to register a permission in the ability builder
      */
     createRegisterFunction(can, options: Record<string, unknown>) {
-      return async (permission: PermissionsTypes.PermissionRule) => {
+      return async (permission: PermissionRule) => {
         const hookContext = createWillRegisterContext({ options, permission });
 
         await state.hooks['before-register.permission'].call(hookContext);
