+++ packages/core/types/src/index.ts
@@ -50,11 +50,9 @@ export {
 };
 
 declare global {
-  // @ts-expect-error - global strapi variable is also defined in the index.d.ts file
   var strapi: LoadedStrapi;
   namespace NodeJS {
     interface Global {
-      // @ts-expect-error - global strapi variable is also defined in the index.d.ts file
       strapi: LoadedStrapi;
     }
   }
